Move the patchers to a separate file
