+++ tests/components/androidtv/test_media_player.py
@@ -1,8 +1,6 @@
 """The tests for the androidtv platform."""
 import logging
-from socket import error as socket_error
 import unittest
-from unittest.mock import patch
 
 from homeassistant.components.androidtv.media_player import (
     AndroidTVDevice,
@@ -10,78 +8,7 @@ from homeassistant.components.androidtv.media_player import (
     setup,
 )
 
-
-def connect_device_success(self, *args, **kwargs):
-    """Return `self`, which will result in the ADB connection being interpreted as available."""
-    return self
-
-
-def connect_device_fail(self, *args, **kwargs):
-    """Raise a socket error."""
-    raise socket_error
-
-
-def adb_shell_python_adb_error(self, cmd):
-    """Raise an error that is among those caught for the Python ADB implementation."""
-    raise AttributeError
-
-
-def adb_shell_adb_server_error(self, cmd):
-    """Raise an error that is among those caught for the ADB server implementation."""
-    raise ConnectionResetError
-
-
-class AdbAvailable:
-    """A class that indicates the ADB connection is available."""
-
-    def shell(self, cmd):
-        """Send an ADB shell command (ADB server implementation)."""
-        return ""
-
-
-class AdbUnavailable:
-    """A class with ADB shell methods that raise errors."""
-
-    def __bool__(self):
-        """Return `False` to indicate that the ADB connection is unavailable."""
-        return False
-
-    def shell(self, cmd):
-        """Raise an error that pertains to the Python ADB implementation."""
-        raise ConnectionResetError
-
-
-PATCH_PYTHON_ADB_CONNECT_SUCCESS = patch(
-    "adb.adb_commands.AdbCommands.ConnectDevice", connect_device_success
-)
-PATCH_PYTHON_ADB_COMMAND_SUCCESS = patch(
-    "adb.adb_commands.AdbCommands.Shell", return_value=""
-)
-PATCH_PYTHON_ADB_CONNECT_FAIL = patch(
-    "adb.adb_commands.AdbCommands.ConnectDevice", connect_device_fail
-)
-PATCH_PYTHON_ADB_COMMAND_FAIL = patch(
-    "adb.adb_commands.AdbCommands.Shell", adb_shell_python_adb_error
-)
-PATCH_PYTHON_ADB_COMMAND_NONE = patch(
-    "adb.adb_commands.AdbCommands.Shell", return_value=None
-)
-
-PATCH_ADB_SERVER_CONNECT_SUCCESS = patch(
-    "adb_messenger.client.Client.device", return_value=AdbAvailable()
-)
-PATCH_ADB_SERVER_AVAILABLE = patch(
-    "androidtv.basetv.BaseTV.available", return_value=True
-)
-PATCH_ADB_SERVER_CONNECT_FAIL = patch(
-    "adb_messenger.client.Client.device", return_value=AdbUnavailable()
-)
-PATCH_ADB_SERVER_COMMAND_FAIL = patch(
-    "{}.AdbAvailable.shell".format(__name__), adb_shell_adb_server_error
-)
-PATCH_ADB_SERVER_COMMAND_NONE = patch(
-    "{}.AdbAvailable.shell".format(__name__), return_value=None
-)
+from . import patchers
 
 
 class TestAndroidTVPythonImplementation(unittest.TestCase):
@@ -89,7 +16,7 @@ class TestAndroidTVPythonImplementation(unittest.TestCase):
 
     def setUp(self):
         """Set up an `AndroidTVDevice` media player."""
-        with PATCH_PYTHON_ADB_CONNECT_SUCCESS, PATCH_PYTHON_ADB_COMMAND_SUCCESS:
+        with patchers.PATCH_PYTHON_ADB_CONNECT_SUCCESS, patchers.PATCH_PYTHON_ADB_COMMAND_SUCCESS:
             aftv = setup("IP:PORT", device_class="androidtv")
             self.aftv = AndroidTVDevice(aftv, "Fake Android TV", {}, None, None)
 
@@ -102,7 +29,7 @@ class TestAndroidTVPythonImplementation(unittest.TestCase):
         https://developers.home-assistant.io/docs/en/integration_quality_scale_index.html
         """
         with self.assertLogs(level=logging.WARNING) as logs:
-            with PATCH_PYTHON_ADB_CONNECT_FAIL, PATCH_PYTHON_ADB_COMMAND_FAIL:
+            with patchers.PATCH_PYTHON_ADB_CONNECT_FAIL, patchers.PATCH_PYTHON_ADB_COMMAND_FAIL:
                 for _ in range(5):
                     self.aftv.update()
                     self.assertFalse(self.aftv.available)
@@ -113,7 +40,7 @@ class TestAndroidTVPythonImplementation(unittest.TestCase):
         assert logs.output[1].startswith("WARNING")
 
         with self.assertLogs(level=logging.DEBUG) as logs:
-            with PATCH_PYTHON_ADB_CONNECT_SUCCESS, PATCH_PYTHON_ADB_COMMAND_SUCCESS:
+            with patchers.PATCH_PYTHON_ADB_CONNECT_SUCCESS, patchers.PATCH_PYTHON_ADB_COMMAND_SUCCESS:
                 # Update 1 will reconnect
                 self.aftv.update()
                 self.assertTrue(self.aftv.available)
@@ -133,12 +60,12 @@ class TestAndroidTVPythonImplementation(unittest.TestCase):
 
         The state should be `None` and the device should be unavailable.
         """
-        with PATCH_PYTHON_ADB_COMMAND_NONE:
+        with patchers.PATCH_PYTHON_ADB_COMMAND_NONE:
             self.aftv.update()
             self.assertFalse(self.aftv.available)
             self.assertIsNone(self.aftv.state)
 
-        with PATCH_PYTHON_ADB_CONNECT_SUCCESS, PATCH_PYTHON_ADB_COMMAND_SUCCESS:
+        with patchers.PATCH_PYTHON_ADB_CONNECT_SUCCESS, patchers.PATCH_PYTHON_ADB_COMMAND_SUCCESS:
             # Update 1 will reconnect
             self.aftv.update()
             self.assertTrue(self.aftv.available)
@@ -154,7 +81,7 @@ class TestAndroidTVServerImplementation(unittest.TestCase):
 
     def setUp(self):
         """Set up an `AndroidTVDevice` media player."""
-        with PATCH_ADB_SERVER_CONNECT_SUCCESS, PATCH_ADB_SERVER_AVAILABLE:
+        with patchers.PATCH_ADB_SERVER_CONNECT_SUCCESS, patchers.PATCH_ADB_SERVER_AVAILABLE:
             aftv = setup(
                 "IP:PORT", adb_server_ip="ADB_SERVER_IP", device_class="androidtv"
             )
@@ -169,7 +96,7 @@ class TestAndroidTVServerImplementation(unittest.TestCase):
         https://developers.home-assistant.io/docs/en/integration_quality_scale_index.html
         """
         with self.assertLogs(level=logging.WARNING) as logs:
-            with PATCH_ADB_SERVER_CONNECT_FAIL, PATCH_ADB_SERVER_COMMAND_FAIL:
+            with patchers.PATCH_ADB_SERVER_CONNECT_FAIL, patchers.PATCH_ADB_SERVER_COMMAND_FAIL:
                 for _ in range(5):
                     self.aftv.update()
                     self.assertFalse(self.aftv.available)
@@ -180,7 +107,7 @@ class TestAndroidTVServerImplementation(unittest.TestCase):
         assert logs.output[1].startswith("WARNING")
 
         with self.assertLogs(level=logging.DEBUG) as logs:
-            with PATCH_ADB_SERVER_CONNECT_SUCCESS:
+            with patchers.PATCH_ADB_SERVER_CONNECT_SUCCESS:
                 self.aftv.update()
                 self.assertTrue(self.aftv.available)
                 self.assertIsNotNone(self.aftv.state)
@@ -199,12 +126,12 @@ class TestAndroidTVServerImplementation(unittest.TestCase):
 
         The state should be `None` and the device should be unavailable.
         """
-        with PATCH_ADB_SERVER_COMMAND_NONE:
+        with patchers.PATCH_ADB_SERVER_COMMAND_NONE:
             self.aftv.update()
             self.assertFalse(self.aftv.available)
             self.assertIsNone(self.aftv.state)
 
-        with PATCH_ADB_SERVER_CONNECT_SUCCESS:
+        with patchers.PATCH_ADB_SERVER_CONNECT_SUCCESS:
             self.aftv.update()
             self.assertTrue(self.aftv.available)
             self.assertIsNotNone(self.aftv.state)
@@ -215,7 +142,7 @@ class TestFireTVPythonImplementation(TestAndroidTVPythonImplementation):
 
     def setUp(self):
         """Set up a `FireTVDevice` media player."""
-        with PATCH_PYTHON_ADB_CONNECT_SUCCESS, PATCH_PYTHON_ADB_COMMAND_SUCCESS:
+        with patchers.PATCH_PYTHON_ADB_CONNECT_SUCCESS, patchers.PATCH_PYTHON_ADB_COMMAND_SUCCESS:
             aftv = setup("IP:PORT", device_class="firetv")
             self.aftv = FireTVDevice(aftv, "Fake Fire TV", {}, True, None, None)
 
@@ -225,7 +152,7 @@ class TestFireTVServerImplementation(TestAndroidTVServerImplementation):
 
     def setUp(self):
         """Set up a `FireTVDevice` media player."""
-        with PATCH_ADB_SERVER_CONNECT_SUCCESS, PATCH_ADB_SERVER_AVAILABLE:
+        with patchers.PATCH_ADB_SERVER_CONNECT_SUCCESS, patchers.PATCH_ADB_SERVER_AVAILABLE:
             aftv = setup(
                 "IP:PORT", adb_server_ip="ADB_SERVER_IP", device_class="firetv"
             )
